actually resolve merge
