+++ code/svb-docker/python/svb_web/core/models.py
@@ -1,8 +1,10 @@
 from django.db import models
 # from datetime import 
 
+
 class IdCardPrintJob(models.Model):
     pass
+
 
 class AccountHolder(models.Model):
     account_holder_number = models.AutoField(primary_key=True)
@@ -14,6 +16,7 @@ class AccountHolder(models.Model):
         null=True,
         on_delete=models.SET_NULL # make referrer field null if the referrer AccountHolder gets deleted
     )
+
 
 class Account(models.Model):
     account_number = models.AutoField(primary_key=True)
@@ -32,3 +35,12 @@ class Account(models.Model):
     last_anchor_event_balance = models.DecimalField(decimal_places=3, max_digits=10) # account balance immediately after last anchor event
 
 
+class NewsAuthor(models.Model):
+    name = models.CharField(max_length=32)  # "Jane Doe"
+    title = models.CharField(max_length=64, blank=True, null=True)  # "Candy QA"
+
+
+class NewsArticle(models.Model):
+    headline = models.CharField(max_length=128)  # "Shrinkflation: King Size Candy 20% Smaller"
+    author = models.ForeignKey('NewsAuthor', on_delete=models.RESTRICT)
+    date_published = models.DateTimeField()
