+++ code/svb-docker/python/svb_web/core/templates/base.html
