+++ code/svb-docker/python/svb_web/core/templates/news_editor.html
@@ -1,4 +1,4 @@
-{% extends "svb_web/base.html" %}
+{% extends "base.html" %}
 {% block content %}
   <!--
     "Editor" Feed (all articles, editor settings shown)
