+++ code/svb-docker/python/svb_web/core/templates/snippets/news_feed.html
