+++ code/svb-docker/python/svb_web/core/urls.py
@@ -5,4 +5,5 @@ from . import views
 
 urlpatterns = [
     path('account/', views.account, name="account"),
+    path('news_editor/', views.news_editor),
 ]
