+++ code/svb-docker/python/svb_web/core/utils/news_feed.py
