+++ code/svb-docker/python/svb_web/svb_web/urls.py
@@ -16,10 +16,9 @@ Including another URLconf
 """
 from django.contrib import admin
 from django.urls import path, include
-from svb_web import views
+from django.contrib.auth import views as auth_views
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include("core.urls")),
-    path('news_editor/', views.news_editor),
 ]
